(BSR) refactor: move venueMap query files
