+++ src/features/venueMap/hook/useGetVenuesInRegion.ts
@@ -2,8 +2,8 @@ import { Venue } from 'features/venue/types'
 import { GeolocatedVenue } from 'features/venueMap/components/VenueMapView/types'
 import { calculateRoundRadiusInKilometers } from 'features/venueMap/helpers/calculateDistanceMap'
 import { isGeolocValid } from 'features/venueMap/helpers/isGeolocValid'
-import { useVenuesInRegionQuery } from 'features/venueMap/useVenuesInRegionQuery'
 import { Region } from 'libs/maps/maps'
+import { useVenuesInRegionQuery } from 'queries/useVenuesInRegionQuery/useVenuesInRegionQuery'
 
 const select = (data: Venue[] | undefined) => {
   return data?.filter(
